Move the parser for person to the persons package
