+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -32,6 +32,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.persons.AddCommand;
+import seedu.address.logic.parser.persons.AddCommandParser;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
