+++ src/test/java/seedu/address/logic/parser/DeleteCommandParserTest.java
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.persons.DeleteCommand;
+import seedu.address.logic.parser.persons.DeleteCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
