+++ src/test/java/seedu/address/logic/parser/EditCommandParserTest.java
@@ -33,6 +33,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.persons.EditCommand;
 import seedu.address.logic.commands.persons.EditCommand.EditPersonDescriptor;
+import seedu.address.logic.parser.persons.EditCommandParser;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
